// projects/frontend/src/api/StrategyType.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type StrategyType =
  | "winner"
  | "goals"
  | "cup_winner"
  | "game_to_penalty"
  | "first_red_card"
  | "first_yellow_card"
  | "penalty_goals"
  | "group_ranking"
  | "round_32_qualifiers"
  | "round_16_qualifiers"
  | "round_8_qualifiers"
  | "round_4_qualifiers"
  | "third_qualifiers"
  | "final";